+++ rl_games/algos_torch/models.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn.functional as F
 from rl_games.algos_torch.torch_ext import CategoricalMasked
 from torch.distributions import Categorical
-from torch import distributions as pyd
+from rl_games.algos_torch.sac_helper import SquashedNormal
 
 
 class BaseModel():
@@ -238,51 +238,6 @@ class ModelA2CContinuousLogStd(BaseModel):
                 + logstd.sum(dim=-1)
 
 class ModelSACContinuous(BaseModel):
-    class TanhTransform(pyd.transforms.Transform):
-        domain = pyd.constraints.real
-        codomain = pyd.constraints.interval(-1.0, 1.0)
-        bijective = True
-        sign = +1
-
-        def __init__(self, cache_size=1):
-            super().__init__(cache_size=cache_size)
-
-        @staticmethod
-        def atanh(x):
-            return 0.5 * (x.log1p() - (-x).log1p())
-
-        def __eq__(self, other):
-            return isinstance(other, TanhTransform)
-
-        def _call(self, x):
-            return x.tanh()
-
-        def _inverse(self, y):
-            # We do not clamp to the boundary here as it may degrade the performance of certain algorithms.
-            # one should use `cache_size=1` instead
-            return self.atanh(y)
-
-        def log_abs_det_jacobian(self, x, y):
-            # We use a formula that is more numerically stable, see details in the following link
-            # https://github.com/tensorflow/probability/commit/ef6bb176e0ebd1cf6e25c6b5cecdd2428c22963f#diff-e120f70e92e6741bca649f04fcd907b7
-            return 2. * (math.log(2.) - x - F.softplus(-2. * x))
-
-
-    class SquashedNormal(pyd.transformed_distribution.TransformedDistribution):
-        def __init__(self, loc, scale):
-            self.loc = loc
-            self.scale = scale
-
-            self.base_dist = pyd.Normal(loc, scale)
-            transforms = [TanhTransform()]
-            super().__init__(self.base_dist, transforms)
-
-        @property
-        def mean(self):
-            mu = self.loc
-            for tr in self.transforms:
-                mu = tr(mu)
-            return mu
 
     def __init__(self, network):
         BaseModel.__init__(self)
@@ -290,8 +245,6 @@ class ModelSACContinuous(BaseModel):
 
     def build(self, config):
         return ModelSACContinuous.Network(self.network_builder.build('sac', **config))
-
-    
     
     class Network(nn.Module):
         def __init__(self, sac_network):
