+++ rl_games/algos_torch/sac_helper.py
@@ -1,0 +1,100 @@
+from rl_games.algos_torch.network_builder import NetworkBuilder
+from torch import distributions as pyd
+import torch
+import torch.nn as nn
+
+class TanhTransform(pyd.transforms.Transform):
+        domain = pyd.constraints.real
+        codomain = pyd.constraints.interval(-1.0, 1.0)
+        bijective = True
+        sign = +1
+
+        def __init__(self, cache_size=1):
+            super().__init__(cache_size=cache_size)
+
+        @staticmethod
+        def atanh(x):
+            return 0.5 * (x.log1p() - (-x).log1p())
+
+        def __eq__(self, other):
+            return isinstance(other, TanhTransform)
+
+        def _call(self, x):
+            return x.tanh()
+
+        def _inverse(self, y):
+            # We do not clamp to the boundary here as it may degrade the performance of certain algorithms.
+            # one should use `cache_size=1` instead
+            return self.atanh(y)
+
+        def log_abs_det_jacobian(self, x, y):
+            # We use a formula that is more numerically stable, see details in the following link
+            # https://github.com/tensorflow/probability/commit/ef6bb176e0ebd1cf6e25c6b5cecdd2428c22963f#diff-e120f70e92e6741bca649f04fcd907b7
+            return 2. * (math.log(2.) - x - F.softplus(-2. * x))
+
+class SquashedNormal(pyd.transformed_distribution.TransformedDistribution):
+        def __init__(self, loc, scale):
+            self.loc = loc
+            self.scale = scale
+
+            self.base_dist = pyd.Normal(loc, scale)
+            transforms = [TanhTransform()]
+            super().__init__(self.base_dist, transforms)
+
+        @property
+        def mean(self):
+            mu = self.loc
+            for tr in self.transforms:
+                mu = tr(mu)
+            return mu
+
+
+class DiagGaussianActor(NetworkBuilder.BaseNetwork):
+    """torch.distributions implementation of an diagonal Gaussian policy."""
+    def __init__(self, output_dim, log_std_bounds, **mlp_args):
+        super().__init__()
+
+        self.log_std_bounds = log_std_bounds
+
+        self.trunk = self._build_mlp(**mlp_args)
+        last_layer = list(self.trunk.children())[-2].out_features
+        self.trunk = nn.Sequential(*list(self.trunk.children()), nn.Linear(last_layer, output_dim))
+
+    def forward(self, obs):
+        mu, log_std = self.trunk(obs).chunk(2, dim=-1)
+
+        # constrain log_std inside [log_std_min, log_std_max]
+        log_std = torch.tanh(log_std)
+        log_std_min, log_std_max = self.log_std_bounds
+        log_std = log_std_min + 0.5 * (log_std_max - log_std_min) * (log_std +
+                                                                     1)
+
+        std = log_std.exp()
+
+        # TODO: Refactor
+        dist = SquashedNormal(mu, std)
+        # Modify to only return mu and std
+        return dist
+
+class DoubleQCritic(NetworkBuilder.BaseNetwork):
+    """Critic network, employes double Q-learning."""
+    def __init__(self, output_dim, **mlp_args):
+        super().__init__()
+
+        self.Q1 = self._build_mlp(**mlp_args)
+        last_layer = list(self.Q1.children())[-2].out_features
+        self.Q1 = nn.Sequential(*list(self.Q1.children()), nn.Linear(last_layer, output_dim))
+
+        self.Q2 = self._build_mlp(**mlp_args)
+        last_layer = list(self.Q2.children())[-2].out_features
+        self.Q2 = nn.Sequential(*list(self.Q2.children()), nn.Linear(last_layer, output_dim))
+
+
+    def forward(self, obs, action):
+        assert obs.size(0) == action.size(0)
+
+        obs_action = torch.cat([obs, action], dim=-1)
+        q1 = self.Q1(obs_action)
+        q2 = self.Q2(obs_action)
+
+        return q1, q2
